+++ client/src/utils/AuthUtil.jsx
@@ -1,7 +1,8 @@
 import axios from 'axios'
 import { loginUser, clearUser } from '../storage/userSlice.js'
+import CryptoJS from 'crypto-js'
 
-export const requestAuth = async (mode, method, value, dispatch, navigate, setErrorMsg, setLoading) => {
+export const requestAuth = async (mode, method, value, dispatch, navigate, setErrorMsg, setLoading, locations=[]) => {
     const url = '/api/auth/' + mode 
     setLoading(true)
     try {
@@ -10,10 +11,11 @@ export const requestAuth = async (mode, method, value, dispatch, navigate, setEr
             if (mode === 'login') {
                 dispatch(loginUser(res.data)) 
                 if (res.data.where.attend) {
-                    navigate('/', {state : {location: value.location, where: res.data.where}}) }
-                else if (res.data.where.isMobile === 'O') {
-                    const resAttend = await axios.post('/api/auth/setAttend', {location: {latitude: value.location.latitude + res.data.hash, longitude: value.location.longitude - res.data.hash}})
-                    navigate('/', {state : {location: value.location, where: resAttend.data.where}})
+                    navigate('/', {state : {location: locations[1], where: res.data.where}})
+                } else if (res.data.where.isMobile === 'O') {
+                    const ciphertext = CryptoJS.AES.encrypt(JSON.stringify(locations), res.data.hash.toString()).toString()
+                    const resAttend = await axios.post('/api/auth/setAttend', {locations: ciphertext})
+                    navigate('/', {state : {location: locations[1], where: resAttend.data.where}})
                 } else {
                     navigate('/attend')
                 }